executor: extract podExecCore and buildSandboxPodSpec from k8s executor
A forthcoming Agent Sandbox CRD executor provisions its sandbox Pod via a
custom resource rather than a direct Pod create, but then execs into the
Pod identically over the pods/exec subresource. Hanging the exec and
file-I/O machinery directly off K8sExecutor made it impossible to share
without duplicating ResolvePath/ReadFile/WriteFile/ListDirectory/Exec and
the streaming cap buffer.

Lift that machinery onto an unexported podExecCore struct and embed it in
K8sExecutor so the exec methods are promoted unchanged; lifecycle (Close,
NetworkPolicy bookkeeping) stays on K8sExecutor. Factor the hardened Pod
spec out of NewK8sExecutor into buildSandboxPodSpec so both executors emit
a byte-identical container. Pure refactor: no behaviour, spec, or error
message changes.
